binbcast: move to a separate file
